% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterloop_v1.R
\name{iCCL}
\alias{iCCL}
\title{iCCL1}
\usage{
iCCL(SeuratObject, min.dim, max.dim, projectname)
}
\arguments{
\item{projectname}{}
}
\value{
cluster plots for dimesnion range between min.dim and max.dim
}
\description{
iCCL1
}
\examples{
iCCL(mySeuratObject, 20, 25, "Macrophages")
}
