% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterloop_v1.R
\name{iCCL}
\alias{iCCL}
\title{iCCL1}
\usage{
iCCL(SeuratObject, min.dim, max.dim)
}
\arguments{
\item{SeuratObject}{your Seurat Objeect}

\item{min.dim}{minimum dimension}

\item{max.dim}{maximum dimenson}
}
\value{
cluster plots for dimesnion range between min.dim and max.dim
}
\description{
iCCL1
}
\examples{
iCCL(mySeuratObject, 20, 25)
}
